+++ src/configs/vue.ts
@@ -425,143 +425,147 @@ export const vueUnConfig: UnConfigFn<
 
   configBuilder
     ?.addConfig(['vue', {includeDefaultFilesAndIgnores: true, filesFallback: DEFAULT_VUE_FILES}])
-    /* Category: Base */
+    .markCategory('Base')
     .addRule('comment-directive', ERROR, [
       // false by default
       {reportUnusedDisableDirectives},
-    ]) // 3️⃣2️⃣ >=4.1.0
-    .addRule('jsx-uses-vars', ERROR) // 3️⃣2️⃣ >=2.0.0
-    /* Category: Priority A: Essential */
-    .addRule('multi-word-component-names', ERROR) // 3️⃣2️⃣ >=7.20.0
-    .addRule('no-arrow-functions-in-watch', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('no-async-in-computed-properties', ERROR) // 3️⃣2️⃣ >=3.8.0
-    .addRule('no-child-content', ERROR) // 3️⃣2️⃣ >=8.1.0
-    .addRule('no-computed-properties-in-data', ERROR) // 3️⃣2️⃣ >=7.20.0
-    .addRule('no-custom-modifiers-on-v-model', vue2Severity(ERROR)) // 2️⃣ >=7.0.0
-    .addRule('no-multiple-template-root', vue2Severity(ERROR)) // 2️⃣ >=7.0.0
-    .addRule('no-deprecated-data-object-declaration', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-destroyed-lifecycle', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-dollar-listeners-api', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-dollar-scopedslots-api', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-events-api', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-filter', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-functional-template', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-html-element-is', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-inline-template', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-props-default-this', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-router-link-tag-prop', vue3Severity(ERROR)) // 3️⃣ >=7.20.0
-    .addRule('no-deprecated-scope-attribute', isLess2_5 ? OFF : ERROR) // 3️⃣ deprecated in 2.5.0 >=6.0.0
-    .addRule('no-deprecated-slot-attribute', isLess2_6 ? OFF : ERROR) // 3️⃣ deprecated in 2.6.0 >=6.1.0
-    .addRule('no-deprecated-slot-scope-attribute', isLess2_6 ? OFF : ERROR) // 3️⃣ deprecated in 2.6.0 >=6.1.0
-    .addRule('no-deprecated-v-bind-sync', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-v-is', isLess3_1 ? OFF : ERROR) // 3️⃣ deprecated in 3.1.0 >=7.11.0
-    .addRule('no-deprecated-v-on-native-modifier', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-v-on-number-modifiers', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-deprecated-vue-config-keycodes', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-dupe-keys', ERROR) // 3️⃣2️⃣ >=3.9.0
-    .addRule('no-dupe-v-else-if', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('no-duplicate-attributes', ERROR) // 3️⃣2️⃣ >=3.0.0
-    .addRule('no-export-in-script-setup', ERROR) // 3️⃣2️⃣ >=7.13.0
-    .addRule('no-expose-after-await', vue3Severity(ERROR)) // 3️⃣ >=8.1.0
-    .addRule('no-lifecycle-after-await', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-mutating-props', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('no-parsing-error', ERROR) // 3️⃣2️⃣ >=3.0.0
-    .addRule('no-ref-as-operand', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('no-reserved-component-names', ERROR) // 3️⃣2️⃣ >=6.1.0
-    .addRule('no-reserved-keys', ERROR) // 3️⃣2️⃣ >=3.9.0
-    .addRule('no-reserved-props', ERROR) // 3️⃣2️⃣ >=8.0.0
-    .addRule('no-shared-component-data', ERROR) // 3️⃣2️⃣ >=3.8.0
-    .addRule('no-side-effects-in-computed-properties', ERROR) // 3️⃣2️⃣ >=3.6.0
-    .addRule('no-template-key', ERROR) // 3️⃣2️⃣ >=3.4.0
-    .addRule('no-textarea-mustache', ERROR) // 3️⃣2️⃣ >=3.0.0
-    .addRule('no-unused-components', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('no-unused-vars', ERROR) // 3️⃣2️⃣ >=3.14.0
-    .addRule('no-use-computed-property-like-method', ERROR) // 3️⃣2️⃣ >=7.15.0
-    .addRule('no-use-v-if-with-v-for', ERROR) // 3️⃣2️⃣ >=4.6.0
-    .addRule('no-useless-template-attributes', ERROR) // 3️⃣2️⃣ >=7.19.0
-    .addRule('no-v-for-template-key-on-child', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('no-v-text-v-html-on-component', ERROR) // 3️⃣2️⃣ >=8.4.0
-    .addRule('no-watch-after-await', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('prefer-import-from-vue', vue3Severity(ERROR)) // 3️⃣ >=8.5.0
-    .addRule('require-component-is', ERROR) // 3️⃣2️⃣ >=3.0.0
-    .addRule('require-prop-type-constructor', ERROR) // 3️⃣2️⃣ >=5.0.0
-    .addRule('require-render-return', ERROR) // 3️⃣2️⃣ >=3.10.0
-    .addRule('require-slots-as-functions', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('require-toggle-inside-transition', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('require-v-for-key', ERROR) // 3️⃣2️⃣ >=3.0.0
-    .addRule('require-valid-default-prop', ERROR) // 3️⃣2️⃣ >=3.13.0
-    .addRule('return-in-computed-property', ERROR) // 3️⃣2️⃣ >=3.7.0
-    .addRule('return-in-emits-validator', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('use-v-on-exact', ERROR) // 3️⃣2️⃣ >=5.0.0
-    .addRule('valid-attribute-name', ERROR) // 3️⃣2️⃣ >=9.0.0
-    .addRule('valid-define-emits', ERROR) // 3️⃣2️⃣ >=7.13.0
-    .addRule('valid-define-props', ERROR) // 3️⃣2️⃣ >=7.13.0
-    .addRule('valid-next-tick', ERROR) // 3️⃣2️⃣ >=7.5.0
-    .addRule('valid-template-root', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-bind', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-cloak', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-else-if', ERROR) //  >=3.11.0
-    .addRule('valid-v-else', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-for', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-html', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-if', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-is', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('valid-v-memo', vue3Severity(ERROR)) // 3️⃣ >=7.16.0
-    .addRule('valid-v-model', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-on', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-once', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-pre', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-show', ERROR) // 3️⃣2️⃣ >=3.11.0
-    .addRule('valid-v-slot', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('valid-v-text', ERROR) // 3️⃣2️⃣ >=3.11.0
-    /* Category: Priority B: Strongly Recommended */
-    .addRule('attribute-hyphenation', ERROR) // 3️⃣2️⃣ >=3.9.0
-    .addRule('component-definition-name-casing', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('first-attribute-linebreak', ERROR) // 3️⃣2️⃣ >=8.0.0
-    .addRule('html-closing-bracket-newline', OFF) // 3️⃣2️⃣🟠 >=4.1.0
-    .addRule('html-closing-bracket-spacing', ERROR) // 3️⃣2️⃣🟠 >=4.1.0
-    .addRule('html-end-tags', ERROR) // 3️⃣2️⃣🟠 >=3.0.0
-    .addRule('html-indent', OFF) // 3️⃣2️⃣🟠 >=3.14.0
-    .addRule('html-quotes', ERROR) // 3️⃣2️⃣🟠 >=3.0.0
-    .addRule('html-self-closing', ERROR, [
-      {
-        html: {
-          // TODO change to `never` once prettier does not `/` to the end of void elements: https://github.com/prettier/prettier/issues/15336
-          void: context.packagesInfo.prettier ? 'any' : 'never',
-          normal: 'never',
-          component: 'never',
+    ]) /** @since 4.1.0 */ // 3️⃣2️⃣
+    .addRule('jsx-uses-vars', ERROR) /** @since 2.0.0 */ // 3️⃣2️⃣
+    .markCategory('Priority A: Essential')
+    .addRule('multi-word-component-names', ERROR) /** @since 7.20.0 */ // 3️⃣2️⃣
+    .addRule('no-arrow-functions-in-watch', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('no-async-in-computed-properties', ERROR) /** @since 3.8.0 */ // 3️⃣2️⃣
+    .addRule('no-child-content', ERROR) /** @since 8.1.0 */ // 3️⃣2️⃣
+    .addRule('no-computed-properties-in-data', ERROR) /** @since 7.20.0 */ // 3️⃣2️⃣
+    .addRule('no-custom-modifiers-on-v-model', vue2Severity(ERROR)) /** @since 7.0.0 */ // 2️⃣
+    .addRule('no-deprecated-data-object-declaration', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-destroyed-lifecycle', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-dollar-listeners-api', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-dollar-scopedslots-api', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-events-api', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-filter', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-functional-template', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-html-element-is', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-inline-template', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-props-default-this', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-router-link-tag-prop', vue3Severity(ERROR)) /** @since 7.20.0 */ // 3️⃣
+    .addRule('no-deprecated-scope-attribute', isLess2_5 ? OFF : ERROR) /** @since 6.0.0 */ // 3️⃣ deprecated in 2.5.0
+    .addRule('no-deprecated-slot-attribute', isLess2_6 ? OFF : ERROR) /** @since 6.1.0 */ // 3️⃣ deprecated in 2.6.0
+    .addRule('no-deprecated-slot-scope-attribute', isLess2_6 ? OFF : ERROR) /** @since 6.1.0 */ // 3️⃣ deprecated in 2.6.0
+    .addRule('no-deprecated-v-bind-sync', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-v-is', isLess3_1 ? OFF : ERROR) /** @since 7.11.0 */ // 3️⃣ deprecated in 3.1.0
+    .addRule('no-deprecated-v-on-native-modifier', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-v-on-number-modifiers', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-deprecated-vue-config-keycodes', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-dupe-keys', ERROR) /** @since 3.9.0 */ // 3️⃣2️⃣
+    .addRule('no-dupe-v-else-if', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('no-duplicate-attributes', ERROR) /** @since 3.0.0 */ // 3️⃣2️⃣
+    .addRule('no-export-in-script-setup', ERROR) /** @since 7.13.0 */ // 3️⃣2️⃣
+    .addRule('no-expose-after-await', vue3Severity(ERROR)) /** @since 8.1.0 */ // 3️⃣
+    .addRule('no-lifecycle-after-await', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-multiple-template-root', vue2Severity(ERROR)) /** @since 7.0.0 */ // 2️⃣
+    .addRule('no-mutating-props', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('no-parsing-error', ERROR) /** @since 3.0.0 */ // 3️⃣2️⃣
+    .addRule('no-ref-as-operand', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('no-reserved-component-names', ERROR) /** @since 6.1.0 */ // 3️⃣2️⃣
+    .addRule('no-reserved-keys', ERROR) /** @since 3.9.0 */ // 3️⃣2️⃣
+    .addRule('no-reserved-props', ERROR) /** @since 8.0.0 */ // 3️⃣2️⃣
+    .addRule('no-shared-component-data', ERROR) /** @since 3.8.0 */ // 3️⃣2️⃣
+    .addRule('no-side-effects-in-computed-properties', ERROR) /** @since 3.6.0 */ // 3️⃣2️⃣
+    .addRule('no-template-key', ERROR) /** @since 3.4.0 */ // 3️⃣2️⃣
+    .addRule('no-textarea-mustache', ERROR) /** @since 3.0.0 */ // 3️⃣2️⃣
+    .addRule('no-unused-components', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('no-unused-vars', ERROR) /** @since 3.14.0 */ // 3️⃣2️⃣
+    .addRule('no-use-computed-property-like-method', ERROR) /** @since 7.15.0 */ // 3️⃣2️⃣
+    .addRule('no-use-v-if-with-v-for', ERROR) /** @since 4.6.0 */ // 3️⃣2️⃣
+    .addRule('no-useless-template-attributes', ERROR) /** @since 7.19.0 */ // 3️⃣2️⃣
+    .addRule('no-v-for-template-key-on-child', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('no-v-text-v-html-on-component', ERROR) /** @since 8.4.0 */ // 3️⃣2️⃣
+    .addRule('no-watch-after-await', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('prefer-import-from-vue', vue3Severity(ERROR)) /** @since 8.5.0 */ // 3️⃣
+    .addRule('require-component-is', ERROR) /** @since 3.0.0 */ // 3️⃣2️⃣
+    .addRule('require-prop-type-constructor', ERROR) /** @since 5.0.0 */ // 3️⃣2️⃣
+    .addRule('require-render-return', ERROR) /** @since 3.10.0 */ // 3️⃣2️⃣
+    .addRule('require-slots-as-functions', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('require-toggle-inside-transition', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('require-v-for-key', ERROR) /** @since 3.0.0 */ // 3️⃣2️⃣
+    .addRule('require-valid-default-prop', ERROR) /** @since 3.13.0 */ // 3️⃣2️⃣
+    .addRule('return-in-computed-property', ERROR) /** @since 3.7.0 */ // 3️⃣2️⃣
+    .addRule('return-in-emits-validator', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('use-v-on-exact', ERROR) /** @since 5.0.0 */ // 3️⃣2️⃣
+    .addRule('valid-attribute-name', ERROR) /** @since 9.0.0 */ // 3️⃣2️⃣
+    .addRule('valid-define-emits', ERROR) /** @since 7.13.0 */ // 3️⃣2️⃣
+    .addRule('valid-define-props', ERROR) /** @since 7.13.0 */ // 3️⃣2️⃣
+    .addRule('valid-next-tick', ERROR) /** @since 7.5.0 */ // 3️⃣2️⃣
+    .addRule('valid-template-root', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-bind', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-cloak', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-else', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-else-if', ERROR) /** @since 3.11.0 */
+    .addRule('valid-v-for', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-html', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-if', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-is', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('valid-v-memo', vue3Severity(ERROR)) /** @since 7.16.0 */ // 3️⃣
+    .addRule('valid-v-model', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-on', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-once', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-pre', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-show', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-slot', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('valid-v-text', ERROR) /** @since 3.11.0 */ // 3️⃣2️⃣
+    .markCategory('Priority B: Strongly Recommended')
+    .addRule('attribute-hyphenation', ERROR) /** @since 3.9.0 */ // 3️⃣2️⃣
+    .addRule('component-definition-name-casing', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('first-attribute-linebreak', ERROR) /** @since 8.0.0 */ // 3️⃣2️⃣
+    .addRule('html-closing-bracket-newline', OFF) /** @since 4.1.0 */ // 3️⃣2️⃣🟠
+    .addRule('html-closing-bracket-spacing', ERROR) /** @since 4.1.0 */ // 3️⃣2️⃣🟠
+    .addRule('html-end-tags', ERROR) /** @since 3.0.0 */ // 3️⃣2️⃣🟠
+    .addRule('html-indent', OFF) /** @since 3.14.0 */ // 3️⃣2️⃣🟠
+    .addRule('html-quotes', ERROR) /** @since 3.0.0 */ // 3️⃣2️⃣🟠
+    .addRule(
+      'html-self-closing',
+      ERROR,
+      [
+        {
+          html: {
+            // TODO change to `never` once prettier does not `/` to the end of void elements: https://github.com/prettier/prettier/issues/15336
+            void: context.packagesInfo.prettier ? 'any' : 'never',
+            normal: 'never',
+            component: 'never',
+          },
         },
-      },
-    ]) // 3️⃣2️⃣🟠 >=3.11.0
-    .addRule('max-attributes-per-line', OFF) // 3️⃣2️⃣🟠 >=3.12.0
-    .addRule('multiline-html-element-content-newline', OFF) // 3️⃣2️⃣🟠 >=5.0.0
-    .addRule('mustache-interpolation-spacing', ERROR) // 3️⃣2️⃣🟠 >=3.13.0
-    .addRule('no-multi-spaces', ERROR) // 3️⃣2️⃣🟠 >=3.12.0
-    .addRule('no-spaces-around-equal-signs-in-attribute', ERROR) // 3️⃣2️⃣🟠 >=5.0.0
-    .addRule('no-template-shadow', ERROR) // 3️⃣2️⃣ >=5.0.0
-    .addRule('one-component-per-file', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('prop-name-casing', ERROR) // 3️⃣2️⃣ >=4.3.0
-    .addRule('require-default-prop', OFF) // 3️⃣2️⃣ >=3.13.0
-    .addRule('require-explicit-emits', vue3Severity(ERROR)) // 3️⃣ >=7.0.0
-    .addRule('require-prop-types', ERROR) // 3️⃣2️⃣ >=3.9.0
-    .addRule('singleline-html-element-content-newline', OFF) // 3️⃣2️⃣🟠 >=5.0.0
+      ] /** @since 3.11.0 */,
+    ) // 3️⃣2️⃣🟠
+    .addRule('max-attributes-per-line', OFF) /** @since 3.12.0 */ // 3️⃣2️⃣🟠
+    .addRule('multiline-html-element-content-newline', OFF) /** @since 5.0.0 */ // 3️⃣2️⃣🟠
+    .addRule('mustache-interpolation-spacing', ERROR) /** @since 3.13.0 */ // 3️⃣2️⃣🟠
+    .addRule('no-multi-spaces', ERROR) /** @since 3.12.0 */ // 3️⃣2️⃣🟠
+    .addRule('no-spaces-around-equal-signs-in-attribute', ERROR) /** @since 5.0.0 */ // 3️⃣2️⃣🟠
+    .addRule('no-template-shadow', ERROR) /** @since 5.0.0 */ // 3️⃣2️⃣
+    .addRule('one-component-per-file', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('prop-name-casing', ERROR) /** @since 4.3.0 */ // 3️⃣2️⃣
+    .addRule('require-default-prop', OFF) /** @since 3.13.0 */ // 3️⃣2️⃣
+    .addRule('require-explicit-emits', vue3Severity(ERROR)) /** @since 7.0.0 */ // 3️⃣
+    .addRule('require-prop-types', ERROR) /** @since 3.9.0 */ // 3️⃣2️⃣
+    .addRule('singleline-html-element-content-newline', OFF) /** @since 5.0.0 */ // 3️⃣2️⃣🟠
     .addRule('v-bind-style', ERROR, [
       'shorthand',
       {
         ...(isMin3_4 && {sameNameShorthand: 'always'}),
       },
-    ]) // 3️⃣2️⃣ >=3.0.0
-    .addRule('v-on-event-hyphenation', vue3Severity(ERROR)) // 3️⃣ >=7.4.0
-    .addRule('v-on-style', ERROR) // 3️⃣2️⃣ >=3.0.0
-    .addRule('v-slot-style', ERROR) // 3️⃣2️⃣ >=6.0.0
-    /* Category: Priority C: Recommended */
-    .addRule('attributes-order', ERROR) // 3️⃣2️⃣ >=4.3.0
-    .addRule('no-lone-template', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('no-multiple-slot-args', ERROR) // 3️⃣2️⃣ >=7.0.0
-    .addRule('no-v-html', ERROR) // 3️⃣2️⃣ >=4.7.0
-    .addRule('order-in-components', ERROR) // 3️⃣2️⃣ >=3.2.0
-    .addRule('this-in-template', ERROR) // 3️⃣2️⃣ >=3.13.0
-    /* Category: Uncategorized */
+    ]) /** @since 3.0.0 */ // 3️⃣2️⃣
+    .addRule('v-on-event-hyphenation', vue3Severity(ERROR)) /** @since 7.4.0 */ // 3️⃣
+    .addRule('v-on-style', ERROR) /** @since 3.0.0 */ // 3️⃣2️⃣
+    .addRule('v-slot-style', ERROR) /** @since 6.0.0 */ // 3️⃣2️⃣
+    .markCategory('Priority C: Recommended')
+    .addRule('attributes-order', ERROR) /** @since 4.3.0 */ // 3️⃣2️⃣
+    .addRule('no-lone-template', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('no-multiple-slot-args', ERROR) /** @since 7.0.0 */ // 3️⃣2️⃣
+    .addRule('no-v-html', ERROR) /** @since 4.7.0 */ // 3️⃣2️⃣
+    .addRule('order-in-components', ERROR) /** @since 3.2.0 */ // 3️⃣2️⃣
+    .addRule('this-in-template', ERROR) /** @since 3.13.0 */ // 3️⃣2️⃣
+    .markCategory('Uncategorized')
     .addRule('block-lang', ERROR, [
       {
         script: {
@@ -569,7 +573,7 @@ export const vueUnConfig: UnConfigFn<
           ...(enforceTypescriptInScriptSection === false && {allowNoLang: true}),
         },
       },
-    ]) // >=7.15.0
+    ]) /** @since 7.15.0 */
     .addRule('block-order', ERROR, [
       {
         order: [
@@ -582,27 +586,31 @@ export const vueUnConfig: UnConfigFn<
           'style[scoped]',
         ],
       },
-    ]) // 3️⃣2️⃣ >=9.16.0
-    .addRule('block-tag-newline', OFF) // 🟠 >=7.1.0
+    ]) /** @since 9.16.0 */ // 3️⃣2️⃣
+    .addRule('block-tag-newline', OFF) /** @since 7.1.0 */ // 🟠
     .addRule('component-api-style', enforceApiStyle == null ? OFF : ERROR, [
       [
         enforceApiStyle === 'setup' ? 'script-setup' : 'options',
         // allows Composition API (not <script setup>)
         isVue2 ? 'composition-vue2' : 'composition',
       ],
-    ]) // >=7.18.0
+    ]) /** @since 7.18.0 */
     .addRule('component-name-in-template-casing', ERROR, [
       'kebab-case',
       {
         registeredComponentsOnly: false,
         ignores: ['/^[A-Z][a-z]+$/' /* Single word components must start with a capital letter */],
       },
-    ]) // >=5.0.0
-    .addRule('component-options-name-casing', ERROR, ['PascalCase' /* default */]) // >=8.2.0
-    .addRule('custom-event-name-casing', ERROR, ['kebab-case' /* default is `camelCase` */]) // >=7.0.0
+    ]) /** @since 5.0.0 */
+    .addRule('component-options-name-casing', ERROR, [
+      'PascalCase' /* default */,
+    ]) /** @since 8.2.0 */
+    .addRule('custom-event-name-casing', ERROR, [
+      'kebab-case' /* default is `camelCase` */,
+    ]) /** @since 7.0.0 */
     .addRule('define-emits-declaration', ERROR, [
       isMin3_3 ? 'type-literal' /* shorter syntax */ : 'type-based' /* default */,
-    ]) // >=9.5.0
+    ]) /** @since 9.5.0 */
     .addRule('define-macros-order', ERROR, [
       {
         order: [
@@ -618,39 +626,39 @@ export const vueUnConfig: UnConfigFn<
         ],
         ...(isMin3_4 && {defineExposeLast: true}),
       },
-    ]) // >=8.7.0
-    .addRule('define-props-declaration', ERROR, [enforcePropsDeclarationStyle]) // >=9.5.0
-    .addRule('define-props-destructuring', ERROR, [{destructure: 'never'}]) // >=10.1.0
-    .addRule('enforce-style-attribute', OFF) // >=9.20.0
-    .addRule('html-button-has-type', ERROR) // >=7.6.0
-    .addRule('html-comment-content-newline', OFF) // >=7.0.0
-    .addRule('html-comment-content-spacing', OFF) // >=7.0.0
-    .addRule('html-comment-indent', OFF) // >=7.0.0
-    .addRule('match-component-file-name', OFF) // >=5.2.0
-    .addRule('match-component-import-name', OFF) // >=8.7.0
-    .addRule('max-lines-per-block', OFF) // >=9.15.0
-    .addRule('max-props', OFF) // >=9.28.0
-    .addRule('max-template-depth', OFF) // >=9.28.0
-    .addRule('new-line-between-multi-line-property', OFF) // >=7.3.0
-    .addRule('next-tick-style', OFF) // >=7.5.0
-    .addRule('no-bare-strings-in-template', OFF) // >=7.0.0
-    .addRule('no-boolean-default', OFF) // >=7.0.0
-    .addRule('no-deprecated-delete-set', vue3Severity(ERROR)) // 3️⃣ >=9.29.0
-    .addRule('no-deprecated-model-definition', vue3Severity(ERROR)) // 3️⃣ >=9.16.0
-    .addRule('no-duplicate-attr-inheritance', ERROR) // >=7.0.0
-    .addRule('no-empty-component-block', ERROR) // >=7.0.0
-    .addRule('no-import-compiler-macros', ERROR) // >=10.0.0
-    .addRule('no-multiple-objects-in-class', ERROR) // >=7.0.0
-    .addRule('no-negated-v-if-condition', ERROR) // >=10.4.0
-    .addRule('no-potential-component-option-typo', ERROR) // >=7.0.0
-    .addRule('no-ref-object-reactivity-loss', ERROR) // >=9.17.0
-    .addRule('no-required-prop-with-default', ERROR) // 3️⃣2️⃣ >=9.6.0
-    .addRule('no-restricted-block', OFF) // >=7.4.0
-    .addRule('no-restricted-call-after-await', OFF) // >=7.4.0
-    .addRule('no-restricted-class', OFF) // >=7.19.0
-    .addRule('no-restricted-component-names', OFF) // >=9.15.0
-    .addRule('no-restricted-component-options', OFF) // >=7.0.0
-    .addRule('no-restricted-custom-event', OFF) // >=7.3.0
+    ]) /** @since 8.7.0 */
+    .addRule('define-props-declaration', ERROR, [enforcePropsDeclarationStyle]) /** @since 9.5.0 */
+    .addRule('define-props-destructuring', ERROR, [{destructure: 'never'}]) /** @since 10.1.0 */
+    .addRule('enforce-style-attribute', OFF) /** @since 9.20.0 */
+    .addRule('html-button-has-type', ERROR) /** @since 7.6.0 */
+    .addRule('html-comment-content-newline', OFF) /** @since 7.0.0 */
+    .addRule('html-comment-content-spacing', OFF) /** @since 7.0.0 */
+    .addRule('html-comment-indent', OFF) /** @since 7.0.0 */
+    .addRule('match-component-file-name', OFF) /** @since 5.2.0 */
+    .addRule('match-component-import-name', OFF) /** @since 8.7.0 */
+    .addRule('max-lines-per-block', OFF) /** @since 9.15.0 */
+    .addRule('max-props', OFF) /** @since 9.28.0 */
+    .addRule('max-template-depth', OFF) /** @since 9.28.0 */
+    .addRule('new-line-between-multi-line-property', OFF) /** @since 7.3.0 */
+    .addRule('next-tick-style', OFF) /** @since 7.5.0 */
+    .addRule('no-bare-strings-in-template', OFF) /** @since 7.0.0 */
+    .addRule('no-boolean-default', OFF) /** @since 7.0.0 */
+    .addRule('no-deprecated-delete-set', vue3Severity(ERROR)) /** @since 9.29.0 */ // 3️⃣
+    .addRule('no-deprecated-model-definition', vue3Severity(ERROR)) /** @since 9.16.0 */ // 3️⃣
+    .addRule('no-duplicate-attr-inheritance', ERROR) /** @since 7.0.0 */
+    .addRule('no-empty-component-block', ERROR) /** @since 7.0.0 */
+    .addRule('no-import-compiler-macros', ERROR) /** @since 10.0.0 */
+    .addRule('no-multiple-objects-in-class', ERROR) /** @since 7.0.0 */
+    .addRule('no-negated-v-if-condition', ERROR) /** @since 10.4.0 */
+    .addRule('no-potential-component-option-typo', ERROR) /** @since 7.0.0 */
+    .addRule('no-ref-object-reactivity-loss', ERROR) /** @since 9.17.0 */
+    .addRule('no-required-prop-with-default', ERROR) /** @since 9.6.0 */ // 3️⃣2️⃣
+    .addRule('no-restricted-block', OFF) /** @since 7.4.0 */
+    .addRule('no-restricted-call-after-await', OFF) /** @since 7.4.0 */
+    .addRule('no-restricted-class', OFF) /** @since 7.19.0 */
+    .addRule('no-restricted-component-names', OFF) /** @since 9.15.0 */
+    .addRule('no-restricted-component-options', OFF) /** @since 7.0.0 */
+    .addRule('no-restricted-custom-event', OFF) /** @since 7.3.0 */
     .addRule(
       'no-restricted-html-elements',
       ERROR,
@@ -658,16 +666,16 @@ export const vueUnConfig: UnConfigFn<
         ...noRestrictedHtmlElementsDefault,
         ...optionsResolved.disallowedHtmlTags,
       }),
-    ) // >=8.6.0
-    .addRule('no-restricted-props', OFF) // >=7.3.0
-    .addRule('no-restricted-static-attribute', OFF) // >=7.0.0
-    .addRule('no-restricted-v-bind', OFF) // >=7.0.0
-    .addRule('no-restricted-v-on', OFF) // >=9.21.0
-    .addRule('no-root-v-if', OFF) // >=9.12.0
-    .addRule('no-setup-props-reactivity-loss', ERROR) // >=9.17.0
-    .addRule('no-static-inline-styles', OFF) // >=7.0.0
-    .addRule('no-template-target-blank', OFF) // >=7.0.0
-    .addRule('no-this-in-before-route-enter', ERROR) // >=7.11.0
+    ) /** @since 8.6.0 */
+    .addRule('no-restricted-props', OFF) /** @since 7.3.0 */
+    .addRule('no-restricted-static-attribute', OFF) /** @since 7.0.0 */
+    .addRule('no-restricted-v-bind', OFF) /** @since 7.0.0 */
+    .addRule('no-restricted-v-on', OFF) /** @since 9.21.0 */
+    .addRule('no-root-v-if', OFF) /** @since 9.12.0 */
+    .addRule('no-setup-props-reactivity-loss', ERROR) /** @since 9.17.0 */
+    .addRule('no-static-inline-styles', OFF) /** @since 7.0.0 */
+    .addRule('no-template-target-blank', OFF) /** @since 7.0.0 */
+    .addRule('no-this-in-before-route-enter', ERROR) /** @since 7.11.0 */
     .addRule('no-undef-components', ERROR, [
       {
         ignorePatterns: [
@@ -679,21 +687,21 @@ export const vueUnConfig: UnConfigFn<
           .flat()
           .filter(Boolean),
       },
-    ]) // >=8.4.0
+    ]) /** @since 8.4.0 */
     // TODO enable if script setup is enforced and only in JS?
-    .addRule('no-undef-properties', OFF) // >=7.20.0
+    .addRule('no-undef-properties', OFF) /** @since 7.20.0 */
     .addRule('no-unsupported-features', ERROR, [
       {version: `^${vuePackageInfo?.versions.full || vuePackageMajorVersion}`},
-    ]) // >=6.1.0
-    .addRule('no-unused-emit-declarations', ERROR) // >=9.19.0
-    .addRule('no-unused-properties', ERROR) // >=7.0.0
-    .addRule('no-unused-refs', ERROR) // >=7.9.0
-    .addRule('no-use-v-else-with-v-for', OFF) // >=9.16.0
-    .addRule('no-useless-mustaches', ERROR) // >=7.0.0
-    .addRule('no-useless-v-bind', ERROR) // >=7.0.0
-    .addRule('no-v-text', OFF) // >=7.17.0
-    .addRule('padding-line-between-blocks', ERROR) // >=6.2.0
-    .addRule('padding-line-between-tags', OFF) // >=9.5.0
+    ]) /** @since 6.1.0 */
+    .addRule('no-unused-emit-declarations', ERROR) /** @since 9.19.0 */
+    .addRule('no-unused-properties', ERROR) /** @since 7.0.0 */
+    .addRule('no-unused-refs', ERROR) /** @since 7.9.0 */
+    .addRule('no-use-v-else-with-v-for', OFF) /** @since 9.16.0 */
+    .addRule('no-useless-mustaches', ERROR) /** @since 7.0.0 */
+    .addRule('no-useless-v-bind', ERROR) /** @since 7.0.0 */
+    .addRule('no-v-text', OFF) /** @since 7.17.0 */
+    .addRule('padding-line-between-blocks', ERROR) /** @since 6.2.0 */
+    .addRule('padding-line-between-tags', OFF) /** @since 9.5.0 */
     .addRule('padding-lines-in-component-definition', ERROR, [
       {
         withinOption: {
@@ -701,51 +709,51 @@ export const vueUnConfig: UnConfigFn<
           props: 'ignore',
         },
       },
-    ]) // >=9.9.0
-    .addRule('prefer-define-options', isMin3_3 ? ERROR : OFF) // >=9.13.0
-    .addRule('prefer-prop-type-boolean-first', ERROR) // >=8.6.0
-    .addRule('prefer-separate-static-class', ERROR) // >=8.2.0
-    .addRule('prefer-true-attribute-shorthand', ERROR) // >=8.5.0
-    .addRule('prefer-use-template-ref', preferUseTemplateRef ? ERROR : OFF) // >=9.31.0
-    .addRule('require-default-export', ERROR) // >=9.28.0
-    .addRule('require-direct-export', ERROR) // >=9.28.0
-    .addRule('require-emit-validator', OFF) // >=7.10.0
-    .addRule('require-explicit-slots', isMin3_3 ? ERROR : OFF) // >=9.21.0
-    .addRule('require-expose', OFF) // >=7.14.0
-    .addRule('require-macro-variable-name', ERROR) // >=9.15.0
-    .addRule('require-name-property', OFF) // >=6.1.0
-    .addRule('require-prop-comment', OFF) // >=9.8.0
-    .addRule('require-typed-object-prop', ERROR) // >=9.16.0
-    .addRule('require-typed-ref', ERROR) // >=9.15.0
-    .addRule('restricted-component-names', OFF) // >=9.32.0
-    .addRule('slot-name-casing', ERROR /* `camelCase` is default */) // >=9.32.0
-    .addRule('script-indent', OFF) // 🟠 >=4.2.0
-    .addRule('sort-keys', OFF) // >=6.2.0
-    .addRule('static-class-names-order', OFF) // >=6.1.0
-    .addRule('v-for-delimiter-style', ERROR, ['in' /* default */]) // >=7.0.0
+    ]) /** @since 9.9.0 */
+    .addRule('prefer-define-options', isMin3_3 ? ERROR : OFF) /** @since 9.13.0 */
+    .addRule('prefer-prop-type-boolean-first', ERROR) /** @since 8.6.0 */
+    .addRule('prefer-separate-static-class', ERROR) /** @since 8.2.0 */
+    .addRule('prefer-true-attribute-shorthand', ERROR) /** @since 8.5.0 */
+    .addRule('prefer-use-template-ref', preferUseTemplateRef ? ERROR : OFF) /** @since 9.31.0 */
+    .addRule('require-default-export', ERROR) /** @since 9.28.0 */
+    .addRule('require-direct-export', ERROR) /** @since 9.28.0 */
+    .addRule('require-emit-validator', OFF) /** @since 7.10.0 */
+    .addRule('require-explicit-slots', isMin3_3 ? ERROR : OFF) /** @since 9.21.0 */
+    .addRule('require-expose', OFF) /** @since 7.14.0 */
+    .addRule('require-macro-variable-name', ERROR) /** @since 9.15.0 */
+    .addRule('require-name-property', OFF) /** @since 6.1.0 */
+    .addRule('require-prop-comment', OFF) /** @since 9.8.0 */
+    .addRule('require-typed-object-prop', ERROR) /** @since 9.16.0 */
+    .addRule('require-typed-ref', ERROR) /** @since 9.15.0 */
+    .addRule('restricted-component-names', OFF) /** @since 9.32.0 */
+    .addRule('script-indent', OFF) /** @since 4.2.0 */ // 🟠
+    .addRule('slot-name-casing', ERROR /* `camelCase` is default */) /** @since 9.32.0 */
+    .addRule('sort-keys', OFF) /** @since 6.2.0 */
+    .addRule('static-class-names-order', OFF) /** @since 6.1.0 */
+    .addRule('v-for-delimiter-style', ERROR, ['in' /* default */]) /** @since 7.0.0 */
     // This rule is not required in Vue 3, as the key is automatically assigned to the elements.
     .addRule('v-if-else-key', vue2Severity(ERROR))
     // TODO change to `[inline, inline-function]` once this is landed: https://github.com/vuejs/eslint-plugin-vue/issues/2460
-    .addRule('v-on-handler-style', ERROR, ['inline']) // >=9.7.0
-    .addRule('valid-define-options', isMin3_3 ? ERROR : OFF) // 3️⃣ >=9.13.0
-    /* Category: Extension Rules */
-    .addRule('array-bracket-newline', OFF) // 🟠 >=7.1.0
-    .addRule('array-bracket-spacing', OFF) // 🟠 >=5.2.0
-    .addRule('array-element-newline', OFF) // 🟠 >=9.9.0
-    .addRule('arrow-spacing', OFF) // 🟠 >=5.2.0
-    .addRule('block-spacing', OFF) // 🟠 >=5.2.0
-    .addRule('brace-style', OFF) // 🟠 >=5.2.0
+    .addRule('v-on-handler-style', ERROR, ['inline']) /** @since 9.7.0 */
+    .addRule('valid-define-options', isMin3_3 ? ERROR : OFF) /** @since 9.13.0 */ // 3️⃣
+    .markCategory('Extension Rules')
+    .addRule('array-bracket-newline', OFF) /** @since 7.1.0 */ // 🟠
+    .addRule('array-bracket-spacing', OFF) /** @since 5.2.0 */ // 🟠
+    .addRule('array-element-newline', OFF) /** @since 9.9.0 */ // 🟠
+    .addRule('arrow-spacing', OFF) /** @since 5.2.0 */ // 🟠
+    .addRule('block-spacing', OFF) /** @since 5.2.0 */ // 🟠
+    .addRule('brace-style', OFF) /** @since 5.2.0 */ // 🟠
     .addRule(
       'camelcase',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules.camelcase ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=5.2.0
-    .addRule('comma-dangle', OFF) // 🟠 >=5.2.0
-    .addRule('comma-spacing', OFF) // 🟠 >=7.0.0
-    .addRule('comma-style', OFF) // 🟠 >=7.0.0
-    .addRule('dot-location', OFF) // 🟠 >=6.0.0
+    ) /** @since 5.2.0 */
+    .addRule('comma-dangle', OFF) /** @since 5.2.0 */ // 🟠
+    .addRule('comma-spacing', OFF) /** @since 7.0.0 */ // 🟠
+    .addRule('comma-style', OFF) /** @since 7.0.0 */ // 🟠
+    .addRule('dot-location', OFF) /** @since 6.0.0 */ // 🟠
     .addRule(
       'dot-notation',
       ...getRuleUnSeverityAndOptionsFromEntry(
@@ -755,113 +763,114 @@ export const vueUnConfig: UnConfigFn<
           inheritFromBase ? undefined : [],
         ],
       ),
-    ) // >=7.0.0
+    ) /** @since 7.0.0 */
     .addRule(
       'eqeqeq',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules.eqeqeq ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=5.2.0
-    .addRule('func-call-spacing', OFF) // 🟠 >=7.0.0
-    .addRule('key-spacing', OFF) // 🟠 >=5.2.0
-    .addRule('keyword-spacing', OFF) // 🟠 >=6.0.0
-    .addRule('max-len', OFF) // 🟠 >=6.1.0
-    .addRule('multiline-ternary', OFF) // 🟠 >=9.7.0
-    .addRule('no-console', ERROR) // >=9.15.0 Do not inherit severity and options
+    ) /** @since 5.2.0 */
+    .addRule('func-call-spacing', OFF) /** @since 7.0.0 */ // 🟠
+    .addRule('key-spacing', OFF) /** @since 5.2.0 */ // 🟠
+    .addRule('keyword-spacing', OFF) /** @since 6.0.0 */ // 🟠
+    .addRule('max-len', OFF) /** @since 6.1.0 */ // 🟠
+    .addRule('multiline-ternary', OFF) /** @since 9.7.0 */ // 🟠
+    .addRule('no-console', ERROR) /** @since 9.15.0 */ // Do not inherit severity and options
     .addRule(
       'no-constant-condition',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-constant-condition'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=7.5.0
+    ) /** @since 7.5.0 */
     .addRule(
       'no-empty-pattern',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-empty-pattern'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=6.0.0
-    .addRule('no-extra-parens', OFF) // 🟠 >=7.0.0
+    ) /** @since 6.0.0 */
+    .addRule('no-extra-parens', OFF) /** @since 7.0.0 */ // 🟠
     .addRule(
       'no-implicit-coercion',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-implicit-coercion'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=9.33.0
+    ) /** @since 9.33.0 */
     .addRule(
       'no-irregular-whitespace',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-irregular-whitespace'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=6.1.0
+    ) /** @since 6.1.0 */
     .addRule(
       'no-loss-of-precision',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-loss-of-precision'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=8.0.0
+    ) /** @since 8.0.0 */
     .addRule(
       'no-negated-condition',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-negated-condition'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=10.4.0
+    ) /** @since 10.4.0 */
     .addRule(
       'no-restricted-syntax',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-restricted-syntax'] ?? OFF,
         inheritFromBase ? undefined : [OFF],
       ),
-    ) // >=5.2.0
+    ) /** @since 5.2.0 */
     .addRule(
       'no-sparse-arrays',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-sparse-arrays'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=7.0.0
+    ) /** @since 7.0.0 */
     .addRule(
       'no-useless-concat',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['no-useless-concat'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=7.0.0
-    .addRule('object-curly-newline', OFF) // 🟠 >=7.0.0
-    .addRule('object-curly-spacing', OFF) // 🟠 >=5.2.0
-    .addRule('object-property-newline', OFF) // 🟠 >=7.0.0
+    ) /** @since 7.0.0 */
+    .addRule('object-curly-newline', OFF) /** @since 7.0.0 */ // 🟠
+    .addRule('object-curly-spacing', OFF) /** @since 5.2.0 */ // 🟠
+    .addRule('object-property-newline', OFF) /** @since 7.0.0 */ // 🟠
     .addRule(
       'object-shorthand',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['object-shorthand'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=8.4.0
-    .addRule('operator-linebreak', OFF) // 🟠 >=7.0.0
+    ) /** @since 8.4.0 */
+    .addRule('operator-linebreak', OFF) /** @since 7.0.0 */ // 🟠
     .addRule(
       'prefer-template',
       ...getRuleUnSeverityAndOptionsFromEntry(
         vanillaFinalFlatConfigRules['prefer-template'] ?? ERROR,
         inheritFromBase ? undefined : [ERROR],
       ),
-    ) // >=7.0.0
-    .addRule('quote-props', OFF) // 🟠 >=8.4.0
-    .addRule('space-in-parens', OFF) // 🟠 >=7.0.0
-    .addRule('space-infix-ops', OFF) // 🟠 >=5.2.0
-    .addRule('space-unary-ops', OFF) // 🟠 >=5.2.0
-    .addRule('template-curly-spacing', OFF) // 🟠 >=7.0.0
+    ) /** @since 7.0.0 */
+    .addRule('quote-props', OFF) /** @since 8.4.0 */ // 🟠
+    .addRule('space-in-parens', OFF) /** @since 7.0.0 */ // 🟠
+    .addRule('space-infix-ops', OFF) /** @since 5.2.0 */ // 🟠
+    .addRule('space-unary-ops', OFF) /** @since 5.2.0 */ // 🟠
+    .addRule('template-curly-spacing', OFF) /** @since 7.0.0 */ // 🟠
     // 🔵 Not working great in Vue files
     .disableAnyRule('ts', 'prefer-function-type')
     .disableAnyRule('ts', 'unified-signatures')
     .disableAnyRule('import', 'first') // May be wrong if multiple <script> blocks are present
     .disableAnyRule('import', 'no-default-export')
     .disableAnyRule('', 'no-useless-assignment') // False positives in script setup
+    .enableConfigTesterForPlugin('vue')
     .addOverrides();
 
   const nuxtLayoutsFilesGlob = resolvePathInVueOrNuxtProjectDir('layouts/**/*.vue');
@@ -922,27 +931,29 @@ export const vueUnConfig: UnConfigFn<
         ignoresFallback: optionsResolved.ignores,
       },
     ])
-    .addRule('alt-text', ERROR) // 🟢
-    .addRule('anchor-has-content', ERROR) // 🟢
-    .addRule('aria-props', ERROR) // 🟢
-    .addRule('aria-role', ERROR) // 🟢
-    .addRule('aria-unsupported-elements', ERROR) // 🟢
-    .addRule('click-events-have-key-events', ERROR) // 🟢
-    .addRule('form-control-has-label', ERROR) // 🟢
-    .addRule('heading-has-content', ERROR) // 🟢
-    .addRule('iframe-has-title', ERROR) // 🟢
-    .addRule('interactive-supports-focus', ERROR) // 🟢
-    .addRule('label-has-for', ERROR, [{allowChildren: true}]) // 🟢
-    .addRule('media-has-caption', ERROR) // 🟢
-    .addRule('mouse-events-have-key-events', ERROR) // 🟢
-    .addRule('no-access-key', ERROR) // 🟢
-    .addRule('no-autofocus', ERROR) // 🟢
-    .addRule('no-distracting-elements', ERROR) // 🟢
-    .addRule('no-onchange', OFF) // Deprecated
-    .addRule('no-redundant-roles', ERROR) // 🟢
-    .addRule('no-static-element-interactions', ERROR) // 🟢
-    .addRule('role-has-required-aria-props', ERROR) // 🟢
-    .addRule('tabindex-no-positive', ERROR) // 🟢
+    .addRule('alt-text', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('anchor-has-content', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('aria-props', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('aria-role', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('aria-unsupported-elements', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('click-events-have-key-events', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('form-control-has-label', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('heading-has-content', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('iframe-has-title', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('interactive-supports-focus', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('label-has-for', ERROR, [{allowChildren: true}]) /** @since 0.1.0 */ // 🟢
+    .addRule('media-has-caption', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('mouse-events-have-key-events', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('no-access-key', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('no-aria-hidden-on-focusable', ERROR) /** @since 2.4.0 */
+    .addRule('no-autofocus', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('no-distracting-elements', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('no-redundant-roles', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('no-role-presentation-on-focusable', ERROR) /** @since 2.4.0 */
+    .addRule('no-static-element-interactions', ERROR) /** @since 2.1.0 */ // 🟢
+    .addRule('role-has-required-aria-props', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('tabindex-no-positive', ERROR) /** @since 0.1.0 */ // 🟢
+    .enableConfigTesterForPlugin('vuejs-accessibility')
     .addOverrides();
 
   const configBuilderPinia = createConfigBuilder(context, configPinia, 'pinia');
@@ -958,11 +969,11 @@ export const vueUnConfig: UnConfigFn<
         ignoreMarkdownCodeBlocks: true,
       },
     ])
-    .addRule('never-export-initialized-store', ERROR) // 🟢
-    .addRule('no-duplicate-store-ids', ERROR) // 🟢
-    .addRule('no-return-global-properties', ERROR) // 🟢
-    .addRule('no-store-to-refs-in-store', ERROR) // 🟢
-    .addRule('prefer-single-store-per-file', ERROR) // 🟢
+    .addRule('never-export-initialized-store', ERROR) /** @since 0.2.0 */ // 🟢
+    .addRule('no-duplicate-store-ids', ERROR) /** @since 0.1.13 */ // 🟢
+    .addRule('no-return-global-properties', ERROR) /** @since 0.1.12 */ // 🟢
+    .addRule('no-store-to-refs-in-store', ERROR) /** @since 0.4.0 */ // 🟢
+    .addRule('prefer-single-store-per-file', ERROR) /** @since 0.1.4 */ // 🟢
     .addRule('prefer-use-store-naming-convention', ERROR, [
       {
         checkStoreNameMismatch: true,
@@ -971,8 +982,9 @@ export const vueUnConfig: UnConfigFn<
             ? configPinia.storesNameSuffix
             : DEFAULT_PINIA_STORE_NAME_SUFFIX,
       },
-    ]) // 🟢
-    .addRule('require-setup-store-properties-export', ERROR) // 🟢
+    ]) /** @since 0.1.2 */ // 🟢
+    .addRule('require-setup-store-properties-export', ERROR) /** @since 0.1.0 */ // 🟢
+    .enableConfigTesterForPlugin('pinia')
     .addOverrides();
 
   const optionsI18nResolved = assignDefaults(
@@ -1010,30 +1022,40 @@ export const vueUnConfig: UnConfigFn<
         }),
       },
     )
-    /* Category: Recommended */
-    .addRule('no-deprecated-i18n-component', isMinVueI18nVersion9 ? ERROR : OFF) // 🟢 >=0.11.0
-    .addRule('no-deprecated-i18n-place-attr', isMinVueI18nVersion9 ? ERROR : OFF) // 🟢 >=0.11.0
-    .addRule('no-deprecated-i18n-places-prop', isMinVueI18nVersion9 ? ERROR : OFF) // 🟢 >=0.11.0
-    .addRule('no-deprecated-modulo-syntax', ERROR) // 🟢 >=3.0.0
-    .addRule('no-deprecated-tc', isMinVueI18nVersion10 ? ERROR : OFF) // 🟢 >=3.0.0
-    .addRule('no-deprecated-v-t', isMinVueI18nVersion10 ? ERROR : OFF) // 🟢 >=3.2.0
-    .addRule('no-html-messages', ERROR) // 🟢 >=0.1.0
-    .addRule('no-i18n-t-path-prop', ERROR) // 🟢 >=0.11.0
-    .addRule('no-missing-keys', ERROR) // 🟢 >=0.1.0
-    .addRule('no-raw-text', ERROR) // 🟢 >=0.2.0
-    .addRule('no-v-html', ERROR) // 🟢 >=0.1.0
-    .addRule('valid-message-syntax', ERROR) // 🟢 >=0.10.0
-    /* Category: Best Practices */
-    .addRule('key-format-style', WARNING) // >=0.9.0
-    .addRule('no-duplicate-keys-in-locale', ERROR) // >=0.9.0
-    .addRule('no-dynamic-keys', WARNING) // >=0.1.0
-    .addRule('no-missing-keys-in-other-locales', ERROR) // >=0.10.0
-    .addRule('no-unknown-locale', ERROR) // >=1.3.0
-    .addRule('no-unused-keys', ERROR) // >=0.1.0
-    .addRule('prefer-sfc-lang-attr', ERROR) // >=1.2.0
-    /* Category: Stylistic Issues */
-    .addRule('prefer-linked-key-with-paren', WARNING) // >=0.10.0
-    .addRule('sfc-locale-attr', ERROR) // >=1.3.0
+    .markCategory('Recommended')
+    .addRule(
+      'no-deprecated-i18n-component',
+      isMinVueI18nVersion9 ? ERROR : OFF,
+    ) /** @since 0.11.0 */ // 🟢
+    .addRule(
+      'no-deprecated-i18n-place-attr',
+      isMinVueI18nVersion9 ? ERROR : OFF,
+    ) /** @since 0.11.0 */ // 🟢
+    .addRule(
+      'no-deprecated-i18n-places-prop',
+      isMinVueI18nVersion9 ? ERROR : OFF,
+    ) /** @since 0.11.0 */ // 🟢
+    .addRule('no-deprecated-modulo-syntax', ERROR) /** @since 3.0.0 */ // 🟢
+    .addRule('no-deprecated-tc', isMinVueI18nVersion10 ? ERROR : OFF) /** @since 3.0.0 */ // 🟢
+    .addRule('no-deprecated-v-t', isMinVueI18nVersion10 ? ERROR : OFF) /** @since 3.2.0 */ // 🟢
+    .addRule('no-html-messages', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('no-i18n-t-path-prop', ERROR) /** @since 0.11.0 */ // 🟢
+    .addRule('no-missing-keys', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('no-raw-text', ERROR) /** @since 0.2.0 */ // 🟢
+    .addRule('no-v-html', ERROR) /** @since 0.1.0 */ // 🟢
+    .addRule('valid-message-syntax', ERROR) /** @since 0.10.0 */ // 🟢
+    .markCategory('Best Practices')
+    .addRule('key-format-style', WARNING) /** @since 0.9.0 */
+    .addRule('no-duplicate-keys-in-locale', ERROR) /** @since 0.9.0 */
+    .addRule('no-dynamic-keys', WARNING) /** @since 0.1.0 */
+    .addRule('no-missing-keys-in-other-locales', ERROR) /** @since 0.10.0 */
+    .addRule('no-unknown-locale', ERROR) /** @since 1.3.0 */
+    .addRule('no-unused-keys', ERROR) /** @since 0.1.0 */
+    .addRule('prefer-sfc-lang-attr', ERROR) /** @since 1.2.0 */
+    .markCategory('Stylistic Issues')
+    .addRule('prefer-linked-key-with-paren', WARNING) /** @since 0.10.0 */
+    .addRule('sfc-locale-attr', ERROR) /** @since 1.3.0 */
+    .enableConfigTesterForPlugin('@intlify/vue-i18n')
     .addOverrides();
 
   const optionsScopedCssResolved = assignDefaults(
@@ -1070,20 +1092,21 @@ export const vueUnConfig: UnConfigFn<
           ),
         },
       ],
-    ) // 3️⃣2️⃣
-    .addRule('no-deprecated-deep-combinator', ERROR) // 3️⃣
-    .addRule('no-deprecated-v-enter-v-leave-class', isVue2 ? OFF : ERROR)
-    .addRule('no-parent-of-v-global', ERROR) // 3️⃣
-    .addRule('no-parsing-error', ERROR) // 3️⃣2️⃣
-    .addRule('no-unused-keyframes', ERROR) // 3️⃣2️⃣
-    .addRule('no-unused-selector', ERROR) // 3️⃣2️⃣
-    .addRule('require-selector-used-inside', OFF)
-    .addRule('require-v-deep-argument', ERROR) // 3️⃣
-    .addRule('require-v-global-argument', ERROR) // 3️⃣
-    .addRule('require-v-slotted-argument', ERROR) // 3️⃣
-    .addRule('v-deep-pseudo-style', ERROR) // 3️⃣
-    .addRule('v-global-pseudo-style', ERROR) // 3️⃣
-    .addRule('v-slotted-pseudo-style', ERROR) // 3️⃣
+    ) /** @since 1.2.0 */ // 3️⃣2️⃣
+    .addRule('no-deprecated-deep-combinator', ERROR) /** @since 1.0.0 */ // 3️⃣
+    .addRule('no-deprecated-v-enter-v-leave-class', isVue2 ? OFF : ERROR) /** @since 1.1.0 */
+    .addRule('no-parent-of-v-global', ERROR) /** @since 1.0.0 */ // 3️⃣
+    .addRule('no-parsing-error', ERROR) /** @since 0.0.1 */ // 3️⃣2️⃣
+    .addRule('no-unused-keyframes', ERROR) /** @since 0.1.0 */ // 3️⃣2️⃣
+    .addRule('no-unused-selector', ERROR) /** @since 0.0.1 */ // 3️⃣2️⃣
+    .addRule('require-selector-used-inside', OFF) /** @since 0.0.1 */
+    .addRule('require-v-deep-argument', ERROR) /** @since 1.0.0 */ // 3️⃣
+    .addRule('require-v-global-argument', ERROR) /** @since 1.0.0 */ // 3️⃣
+    .addRule('require-v-slotted-argument', ERROR) /** @since 1.0.0 */ // 3️⃣
+    .addRule('v-deep-pseudo-style', ERROR) /** @since 2.3.0 */ // 3️⃣
+    .addRule('v-global-pseudo-style', ERROR) /** @since 2.3.0 */ // 3️⃣
+    .addRule('v-slotted-pseudo-style', ERROR) /** @since 2.3.0 */ // 3️⃣
+    .enableConfigTesterForPlugin('vue-scoped-css')
     .addOverrides();
 
   return {